add MessageEntity in deleteUser

+++ backend/src/user/user.service.ts
@@ -8,6 +8,7 @@ import { RoomService } from '../chat/room/room.service';
 import { RoomEntity } from 'src/chat/room/entities/room.entity';
 import { UserToRoomEntity } from 'src/chat/room/entities/user.to.room.entity';
 import ConnectedUserEntity from 'src/chat/connected-user/connected-user.entity';
+import { MessageEntity } from 'src/chat/message/message.entity';
 
 @Injectable()
 export class UserService {
@@ -130,6 +131,12 @@ export class UserService {
 			.createQueryBuilder()
 			.delete()
 			.from(ConnectedUserEntity)
+			.where('userId = :userId', { userId })
+			.execute();
+		await getConnection()
+			.createQueryBuilder()
+			.delete()
+			.from(MessageEntity)
 			.where('userId = :userId', { userId })
 			.execute();
 		await getConnection()
